feat(system): sdl2: select display mode closest to 60hz
when selecting a display mode, sdl will use default desktop mode to search for the appropriate refresh rate. This patch set the desired refresh rate to 60Hz. That means if the desktop mode is 30Hz or 120Hz as we have seen on some devices, selected mode for a specific resolution will be the 60Hz one.